chore(auth): rename category level request.ts to input.ts
